change register page, login page, and welcome page to ts
