+++ client/components/App.tsx
@@ -1,34 +1,36 @@
-import React from 'react';
-import { Route, Routes, Link } from 'react-router-dom';
-
-// import pages
-import WelcomePage from '../pages/WelcomePage';
+import React, { useEffect, useState } from 'react';
+import { Routes, Route } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { setCurrentUser } from '../store/userReducer';
 import LoginPage from '../pages/LoginPage';
-import RegisterPage from '../pages/RegisterPage';
+import SignUpPage from '../pages/RegisterPage';
+import WelcomePage from '../pages/WelcomePage';
 
-// import redux hooks and action creators
-import { useAppDispatch, useAppSelector } from '../app/hooks';
-import { increment, incrementByAmount } from '../features/counter/counterSlice';
+function App() {
+  const dispatch = useDispatch();
+  // const currentUser = useSelector((state: any) => state.userReducer.currentUser);
+    
 
-const App = () => {
-  const count = useAppSelector((state) => state.counter.value);
-  const dispatch = useAppDispatch();
-
-  function handleClick() {
-    dispatch(increment());
-    dispatch(incrementByAmount(3));
-  }
+  useEffect(() => {
+    // fetch to server to check for a session cookie
+    fetch('/login/auth')
+      .then(res => res)
+      .then(res => {
+        // if the server returns a 200 status code, set CurrentUser (on State) to the response from the server, which is the logged in user's username
+        if (res.status === 200) dispatch(setCurrentUser(res));
+      })
+      .catch(err => {
+        console.log(err);
+      });
+  }, []);
 
   return (
-    <>
-      <button onClick={handleClick}>Click me</button>
-      <Routes>
-        <Route path="/" element={<WelcomePage />} />
-        <Route path="/login" element={<LoginPage />} />
-        <Route path="/register" element={<RegisterPage />} />
-      </Routes>
-    </>
+    <Routes>
+      <Route path={'/'} element={<WelcomePage/>}></Route>
+      <Route path={'/loginpage'} element={<LoginPage/>}></Route>
+      <Route path={'/signuppage'} element={<SignUpPage/>}></Route>
+    </Routes>
   );
-};
+}
 
 export default App;
