--- client/pages/WelcomePage.jsx
@@ -1,3 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-// will eventually import footer & nav bars
