+++ client/pages/WelcomePage.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
+
 
 const WelcomePage = () => {
   return (
